Adds the prediction columns to a piped command set

Adds a new column with the results form predict_fit() to a piped command set. If add_interval is set to TRUE, then it will add two additional columns, one for the lower and another for the upper prediction interval bounds.

predict_to_column(df, model, add_interval = FALSE, interval = 0.95,
  vars = c("fit", "upper", "lower"))

Arguments

df
A data.frame or tibble

model
An R object, typically an R model, but it can also bee a regular data frame that has a parsed model already loaded.

add_interval
Switch that indicates if the prediction interval columns should be added. Defaults to FALSE

interval
The prediction interval, defaults to 0.95. It is ignored if add_interval is set to FALSE

vars
The name of the variables that this function will produce. It defaults to "fit", "upper", and "lower".

Examples

library(dplyr)

df <- data.frame(x = c(1, 2, 5, 6 ,6), y = c(2, 3, 6, 5, 4))
model <- lm(x ~ y, df)
df %>%
  predict_to_column(model, add_interval = TRUE)

#>   x y fit     upper        lower
#> 1 1 2 1.6  8.007185 -4.807184721
#> 2 2 3 2.8  8.575358 -2.975358261
#> 3 5 6 6.4 12.807185 -0.007184721
#> 4 6 5 5.2 10.975358 -0.575358261
#> 5 6 4 4.0  9.548785 -1.548784735