Tests base predict function against tidypredict

Runs predict() and the predict_to_column() functions and compares the results. Because of decimal precision, some results will be different, but the difference will be very small. That's why a "pass" for a test will be if there is no individual result that differs more than the 'threshold' amount set.

test_predictions(model, df = model$model, threshold = 1e-12,
  include_intervals = FALSE, max_rows = NULL)

Arguments

model
An R model object

df
A data frame that contains all of the needed fields to run the prediction. It defaults to the "model" data frame object inside the model object.

threshold
The number that a given result difference, between predict() and predict_to_column() should not exceed. The default value is 0.000000000001 (1e-12)

include_intervals
Switch to indicate if the prediction intervals should be included in the test. It defaults to FALSE.

max_rows
The number of rows in the object passed in the df argument. Highly recommended for large data sets.

Examples

df <- data.frame(x = c(1, 2, 5, 6 ,6), y = c(2, 3, 6, 5, 4))
model <- lm(x ~ y, df)
test_predictions(model, include_intervals = TRUE)

#> tidypredict test results
#> Difference threshold: 1e-12
#> 
#>  All results are within the difference threshold